requirements.txt
django==2.2.9
gunicorn==20.0.4
mysqlclient==1.4.6
markdown==3.1.1
django-uuslug==1.2.0
django-haystack==2.8.1
whoosh==2.7.4
jieba==0.40
requests==2.22.0
